Generic EvaluationContext to support multiple implementations of evaluation & translation
